pyproject.toml
[project]
dependencies = [
  "django-extensions>=3.2.3", # Adds additional features to Django.
  "django>=5.1.4", # Django Framework.
  "django-environ>=0.11.2", # Allows for environment variables management.
  "djangorestframework>=3.15.2", # Enables building APIs with Django.
  "django-seed>=0.3.1", # Provides data seeding functionality.
  "factory-boy>=3.3.1", # Adds support for factories in tests.
  "django-storages>=1.14.4", # Enables the use of S3 storage.
  "django-import-export>=4.3.3", # Adds data import and export capabilities.
  "drf-spectacular>=0.28.0", # Automatically generates API documentation.
  "django-redis>=5.4.0", # Integrates Redis as a cache backend.
  "channels>=4.2.0", # Enables WebSocket communication.
  "psycopg>=3.2.3", # Provides support for PostgreSQL database.
  "requests>=2.32.3", # Provides request/response monitoring and profiling.
  "ruff>=0.8.3", # Python linter and formatter.
  "pytest-django>=4.9.0", # Facilitates unit testing with Django.
  "boto3>=1.35.91",
  "django-sonar>=0.1.8",
]
description = "This is a practice project where i will replicate netflix like backend system using django."
name = "netflix-clone"
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

[tool.ruff]
# Set the maximum line length to 79 characters.
line-length = 79

lint.ignore = [
  "F403",
  "F405",
  "E501",
] # Ignore specific rules related to unused imports.
